refactor(main): move the Team API wire decoders out of the manager
`remote-server-manager.ts` carried ~53 decoders and three URL builders in 757 lines that contain no
`this.` at all. They move to five files named for the wire area they decode, plus one for the URL and
query shapes the client builds.

`remote-host-decoding.ts` holds what all five share -- `ResponseDecoder`, `decodeVoid` -- and the
`FromHost`/`FromMain` paragraph, which is now a file header rather than a comment sitting above one
decoder. Its twin in `src/preload/index.ts` still points at the old path; the next commit moves both
halves together, because the convention only works while a reader can find the twin.

Pure move. The manager re-exports all 23 previously public names so no importer changes in this
commit; the next one repoints them and deletes the block.
